SLE-918: Refactor "Open in IDE" logic to be re-usable
In order to be able to re-use the logic of the "Open in IDE" feature for "Open fix suggestion" a small refactoring was required.
