+++ org.sonarlint.eclipse.ui/src/org/sonarlint/eclipse/ui/internal/job/OpenIssueInEclipseJob.java
@@ -22,12 +22,8 @@ package org.sonarlint.eclipse.ui.internal.job;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.IJobChangeEvent;
@@ -36,24 +32,15 @@ import org.eclipse.core.runtime.jobs.JobChangeAdapter;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.PartInitException;
-import org.eclipse.ui.PlatformUI;
 import org.sonarlint.eclipse.core.SonarLintLogger;
-import org.sonarlint.eclipse.core.documentation.SonarLintDocumentation;
-import org.sonarlint.eclipse.core.internal.backend.ConfigScopeSynchronizer;
-import org.sonarlint.eclipse.core.internal.backend.SonarLintBackendService;
 import org.sonarlint.eclipse.core.internal.engine.connected.ResolvedBinding;
-import org.sonarlint.eclipse.core.internal.jobs.AnalysisReadyStatusCache;
 import org.sonarlint.eclipse.core.internal.jobs.AnalyzeProjectRequest.FileWithDocument;
 import org.sonarlint.eclipse.core.internal.jobs.AnalyzeProjectsJob;
 import org.sonarlint.eclipse.core.internal.jobs.TaintIssuesMarkerUpdateJob;
 import org.sonarlint.eclipse.core.internal.markers.MarkerMatcher;
-import org.sonarlint.eclipse.core.internal.vcs.VcsService;
 import org.sonarlint.eclipse.core.resource.ISonarLintFile;
 import org.sonarlint.eclipse.core.resource.ISonarLintProject;
-import org.sonarlint.eclipse.ui.internal.SonarLintUiPlugin;
-import org.sonarlint.eclipse.ui.internal.dialog.AwaitProjectConnectionReadyDialog;
 import org.sonarlint.eclipse.ui.internal.preferences.SonarLintPreferencePage;
-import org.sonarlint.eclipse.ui.internal.util.BrowserUtils;
 import org.sonarlint.eclipse.ui.internal.util.MessageDialogUtils;
 import org.sonarlint.eclipse.ui.internal.util.PlatformUtils;
 import org.sonarlint.eclipse.ui.internal.views.RuleDescriptionWebView;
@@ -66,139 +53,36 @@ import org.sonarsource.sonarlint.core.rpc.protocol.client.issue.IssueDetailsDto;
  *  "Open in IDE": After covering most of cases where we cannot match the issue locally, this tries to match with the
  *                 actual markers we have on the files.
  */
-public class OpenIssueInEclipseJob extends Job {
-  @Nullable
-  private ISonarLintFile file;
-
+public class OpenIssueInEclipseJob extends AbstractOpenInEclipseJob {
   private final String name;
   private final IssueDetailsDto issueDetails;
-  private final ISonarLintProject project;
   private final ResolvedBinding binding;
   private final boolean recreatedMarkersAlready;
   private final boolean askedForPreferenceChangeAlready;
 
   public OpenIssueInEclipseJob(OpenIssueContext context) {
-    super(context.getName());
+    super(context.getName(), context.getProject(), false);
 
     this.name = context.getName();
     this.issueDetails = context.getIssueDetails();
-    this.project = context.getProject();
     this.binding = context.getBinding();
     this.recreatedMarkersAlready = context.getRecreatedMarkersAlready();
     this.askedForPreferenceChangeAlready = context.getAskedForPreferenceChangeAlready();
   }
 
   @Override
-  protected IStatus run(IProgressMonitor monitor) {
-    // 1) We have to await the analysis getting ready for this project. When also setting up the connection / binding
-    // with this Open in IDE request it isn't the case. We want to display a nice dialog informing the user that they
-    // have to wait for a few more moments.
-    if (!AnalysisReadyStatusCache.getAnalysisReadiness(ConfigScopeSynchronizer.getConfigScopeId(project))) {
-      // Show the console so the user will see the progress!
-      SonarLintUiPlugin.getDefault().getSonarConsole().bringConsoleToFront();
-
-      // We have to run the dialog in the UI thread, therefore we have to work with references here and synchronized
-      // threads for the job to proceed correctly later.
-      var statusRef = new AtomicReference<IStatus>();
-      var cancelledByJob = new AtomicBoolean();
-      statusRef.set(Status.OK_STATUS);
-      cancelledByJob.set(false);
-      Display.getDefault().syncExec(() -> {
-        var dialog = new AwaitProjectConnectionReadyDialog(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell());
-        new AwaitProjectConnectionReadyJob(project, dialog, 0)
-          .schedule(AwaitProjectConnectionReadyJob.SCHEDULE_TIMER_MS);
-        dialog.open();
-        if (dialog.cancelledByJob() || dialog.cancelledByUser()) {
-          SonarLintLogger.get().debug("Open in IDE after setting up Connected Mode cancelled by "
-            + (dialog.cancelledByUser() ? "the user." : "timeout of the background job."));
-          statusRef.set(Status.CANCEL_STATUS);
-        }
-        cancelledByJob.set(dialog.cancelledByJob());
-      });
-
-      // We only want to show a new dialog when the connection was not yet ready in time and not when the user manually
-      // cancelled it (e.g. because their priority changed, it took too long for them or whatever)
-      var status = statusRef.get();
-      if (Status.CANCEL_STATUS == status) {
-        if (cancelledByJob.get()) {
-          MessageDialogUtils.openInIdeError("The previous dialog was closed either manually by you or the connection was not ready in "
-            + "time. The reason for the latter could be a slow network connection.");
-        }
-        return status;
-      }
-    }
-
-    // 2) Check if matching between remote and local path does work / branches match
-    // INFO: When we re-run this job we don't have to do all the checks again!
-    if (file == null) {
-      var fileOpt = tryGetLocalFile();
-      if (fileOpt.isEmpty()) {
-        return Status.CANCEL_STATUS;
-      }
-      file = fileOpt.get();
-
-      if (!tryMatchBranches()) {
-        return Status.CANCEL_STATUS;
-      }
-    }
-
-    try {
-      // 3) Handle normal issues / Taint Vulnerabilities differently
-      return issueDetails.isTaint() ? handleTaintIssue() : handleNormalIssue();
-    } catch (CoreException e) {
-      var message = "An error occured while trying to match the issue locally.";
-
-      MessageDialogUtils.openInIdeError(message + " Please see the console for the full error log!");
-      SonarLintLogger.get().error(message, e);
-    }
-
-    return Status.CANCEL_STATUS;
+  protected IStatus actualRun() throws CoreException {
+    return issueDetails.isTaint() ? handleTaintIssue() : handleNormalIssue();
   }
 
-  /** File check: We try to convert the server path to IDE path in order to find the correct file */
-  private Optional<ISonarLintFile> tryGetLocalFile() {
-    // Check if file exists in project based on the server to IDE path matching
-    var fileOpt = project.find(issueDetails.getIdeFilePath().toString());
-    if (fileOpt.isEmpty()) {
-      MessageDialogUtils.openInIdeError("The file containing the issue cannot be found in the project '"
-        + project.getName() + "'. Maybe it was already changed locally!");
-      return Optional.empty();
-    }
-
-    return fileOpt;
+  @Override
+  protected String getIdeFilePath() {
+    return issueDetails.getIdeFilePath().toString();
   }
 
-  /** Branch check: Local and remote information should match (if no local branch found, at least try your best) */
-  private boolean tryMatchBranches() {
-    var branch = issueDetails.getBranch();
-
-    // In case SLCORE is not yet ready for the SonarProjectBranchService, we also check the "legacy" and Eclipse-
-    // specific integration with EGit. If no EGit is installed, this will of course will not work, but as it is only
-    // used in case of the Open in IDE with automatic Connected Mode setup.
-    Optional<String> localBranch = Optional.empty();
-    try {
-      var response = SonarLintBackendService.get().getMatchedSonarProjectBranch(project);
-      localBranch = Optional.ofNullable(response.getMatchedSonarProjectBranch());
-    } catch (InterruptedException | ExecutionException err) {
-      SonarLintLogger.get().debug("Cannot get matched branch from backend, trying local VCS service", err);
-    }
-    if (localBranch.isEmpty()) {
-      localBranch = VcsService.getCachedSonarProjectBranch(project);
-    }
-
-    if (localBranch.isEmpty()) {
-      // This error message may be misleading to COBOL / ABAP developers but that is okay for now :>
-      MessageDialogUtils.openInIdeInformation("The local branch of the project '" + project.getName()
-        + "' could not be determined. SonarLint now can only try to find the matching local issue!");
-    } else if (!branch.equals(localBranch.get())) {
-      MessageDialogUtils.openInIdeError("The local branch '" + localBranch.get() + "' of the project '"
-        + project.getName() + "' does not match the remote branch '" + branch + "'. "
-        + "Please checkout the correct branch and invoke 'Open in IDE' once again!");
-      BrowserUtils.openExternalBrowser(SonarLintDocumentation.BRANCH_AWARENESS, Display.getDefault());
-      return false;
-    }
-
-    return true;
+  @Override
+  protected String getBranch() {
+    return issueDetails.getBranch();
   }
 
   /** Handle normal issues: They can be present as On-the-fly / Report markers */
